Rename files to match class names for better consistency
File name changes:
- validators.py → audio_file_validator.py (DefaultAudioFileValidator)
- builders.py → transcription_payload_builder.py (DefaultTranscriptionPayloadBuilder)
- collectors.py → transcription_result_collector.py (DefaultTranscriptionResultCollector)
- savers.py → output_saver.py (DefaultOutputSaver)
- displays.py → result_display.py (DefaultResultDisplay)
- providers.py → transcription_parameter_provider.py (TranscriptionParameterProvider)
- waiters.py → queue_waiter.py (QueueWaiter)
- factories.py → runpod_endpoint_factory.py (DefaultRunPodEndpointFactory)
- client.py → audio_transcription_client.py (AudioTranscriptionClient)

Benefits:
- File names directly correspond to their primary class names
- Better consistency and clarity
- Easier to locate specific functionality
- Follows Python naming conventions
- Improved code navigation

Updated imports in:
- __init__.py: Updated all import statements
- audio_transcription_client.py: Updated internal imports

All 66 unit tests pass
Maintains backward compatibility
No breaking changes to public API
