+++ src/clients/audio_transcription/transcription_payload_builder.py
